requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.5
streamlit==1.28.1
faiss-cpu==1.7.4
pandas==2.1.3
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
numpy==1.24.3
sentence-transformers==2.3.1
huggingface-hub==0.19.4
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
